backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pandas==2.2.3
python-docx==1.1.2
python-multipart==0.0.9
openpyxl==3.1.5
